pyproject.toml
[tool.poetry]
name = "poetrytemplate"
version = "0.1.0"
description = ""
authors = ["x <x@det.nl>"]
readme = "README.md"
repository = "https://github.com/Dynamic-Energy-Trading/poetrytemplate"
packages = [
    { include = "src" }, # Specifies the location of the package
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
python-dotenv = "^1.1.0"
detquantlib = "^3.2.7"

[tool.poetry.group.dev.dependencies]
invoke = "^2.2.0"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
black = "^25.1.0"
darglint = "^1.8.1"
isort = "^6.0.1"
colorama = "^0.4.6"
pymarkdownlnt = "^0.9.30"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
line_length = 99
force_grid_wrap = 0
use_parentheses = true
known_third_party = ["invoke"]

[tool.black]
line-length = 99

[tool.pymarkdown]
plugins.md013.enabled = false # Disable line length requirements
plugins.md040.enabled = false # Disable fenced code blocks language requirements
